Manually link compiler-rt into libc++, libc++abi, libunwind
We are using statically linked compiler-rt builtin library which
works fine in most cases. However, libc++, libc++abi and libunwind
use -nodefaultlibs flag (for legitimate reasons) which avoids
linking in compiler-rt builtins which results in missing
implementation of builtin functions, so we need to link in
compiler-rt manually to avoid this issue.

Change-Id: I77adb1b2453b6fdff4e1bd59a554651555e990ab