Add notion of a 'default' entry in package config that can be overridden
Currently, the boot_dev package config defines a second entry for
apps.config.  If this ends up later on in the manifest entry the later
entry 'wins', but this confuses the installer and is a fragile
practice. This makes specifying multiple entries for the same bootfs
path in one configuration an error and allows adding the "default" key
to a file/resource entry to indicate that the entry is only a default
and can be safely overriden.

Test: Run gen.py -m default,boot_dev and verify only one entry for
apps.config exists in the generated manifest
(out/debug-x86-64/gen/packages/gn/user.bootfs.manifest)

Change-Id: I37b3172d379561d991aa9de204ddd2da5301f234