Put drivers and dependencies into /boot instead of /system
The current version of the device manager doesn't load drivers added to
the system after the device they service is discovered so devices
discovered at boot (say attached via PCI) may not be loaded if they are
loaded from the regular minfs filesystem since this loads later. To work
around this issue until the device manager gets smarter, this arranges
for drivers and their runtime dependencies (like libc++.so.2) to be
loaded into /boot which is mounted before the device manager enumerates
local devices. To achieve this, this adds a new component type to module
configurations called "drivers" which is similar to "binaries" except
that it installs into /boot. This produces a separate manifest which is
passed to mkbootfs along with --target=boot to generate the correct
header.  The installer in //scripts/installer/ also knows about this
manifest file and uses it when producing the installer image.

Change-Id: Ie609b2ec1be11ed647b62a52e0a204ee694242f6