Only expand package() template out in target toolchain
If a build file containing a package definition is referenced in
another toolchain, say by a host tool or a target in the shared library
toolchain depending on zlib, then the package template will attempt to
expand in that toolchain and get very confused since the dependencies
(in that toolchain) will not produce the artifacts that the package
definition expects. This makes the package template only expand out to
something with steps in the target toolchain. The build edges produced
by gen.py are all in the target toolchain. It's still possible to
express dependencies on targets in other toolchains and put the
artifacts those targets produce into the package by referring to their
path relative to the default target toolchain's path (i.e.
x64-shared/libminizip.so).

Change-Id: I6d129a180339bac6ba86ac18f415734b0a5d961f