Always emit manifests, even when empty
This updates the package() template to always emit the archive_manifest,
boot_manifest and system_manifest files, not only when they are non-empty.
This ensures that when when a package changes from having entries for these
manifests to having no entries that the generation step leaves behind an
empty file instead of leaving a file with stale contents.

Change-Id: Idc3f1b00b72684b5ac933f9a464e6e0914a81fc6