Allow libraries to be marked as used by drivers to go into /boot/lib
Drivers can be instantiated from /boot before /system/lib is available,
so libraries used by these drivers need to be placed into /boot/lib
until our driver loading is sophisticated enough to be able to load
drivers later in the boot sequence. This allows tagging a library entry
in a package manifest with "used_by_drivers" to indicate this. This is
needed concretely for the libaudio-dispatcher-pool.so library which is
used by several audio drivers.

Change-Id: I21990cb73eb457e3998872fdf3e7262ec7d2733c