Detect duplicate package names in configs
This detects and raises an error when multiple packages with the same
name are encountered in a build. There are two cases to check - the
same key used multiple times in the same "packages" block and the same
name being used in different "packages" blocks from different files.
The former is detected by hooking the json parser to detect duplicate
keys, since they would usually be folded away by the json parser. The
latter is detected by checking if a package name has been seen before.

Fixes: TO-369
Change-Id: I8a57ff62fbae754d731a36dd034ec3bd1270c9b3