[packages] Allow apps to choose destination
Previously, if app was set to true and a binary specified a dest,
the build would fail due to the app variable not being used (its
use was hidden by a nested if). This change allows an app to choose
a destination within the app hierarchy, which is necessary for
agents.

Change-Id: I819e0b87bfde27b3a57a12d69cd352968724cb61